Unset GNUPGHOME to use NixOS systemd unit
If GNUPGHOME is set gpg takes a hash of the folder and creates a socket
depending on that hash.  This results in gpg not picking up the socket
provided by NixOS.  If ~/.gnupg is a symlink to the desired location
instead the provided socket is used.

This fixes the pinentry issues on NixOS.